+++ src/tonpy/tvm/__init__.py
@@ -1,5 +1,5 @@
 # Copyright (c) 2023 Disintar LLP Licensed under the Apache License Version 2.0
 
 from tonpy.tvm.tvm import TVM, method_name_to_id
-from tonpy.tvm.c7 import BlockId, C7, StepInfo
+from tonpy.tvm.c7 import C7, StepInfo
 from tonpy.tvm.emulator import Emulator