+++ src/tonpy/types/__init__.py
@@ -8,6 +8,8 @@ from tonpy.types.tlb import TLB, RecordBase
 from tonpy.types.tlb_types import RefT, NatWidth, TLBComplex, Int, UInt, Bits, NatLeq, NatLess
 from tonpy.types.stack import StackEntry, Stack, Continuation
 from tonpy.types.keys import *
+from tonpy.types.blockid import *
+from tonpy.types.lite_utils import *
 from tonpy.types.liteclient import *
 from tonpy.types.address import *