+++ src/tonpy/types/tlb.py
@@ -7,13 +7,20 @@ from typing import Optional, List, Union, Callable
 from tonpy.types import CellSlice, CellBuilder, Cell
 
 
-def rec_dump(item):
+def rec_dump(item, dump_bin_as_hex):
     output = {}
     for name in item.field_names:
         value = getattr(item, name)
         if issubclass(type(value), RecordBase):
-            output[name] = rec_dump(value)
+            output[name] = rec_dump(value, dump_bin_as_hex=dump_bin_as_hex)
         else:
+            if dump_bin_as_hex and isinstance(value, str) and len(value) % 8 == 0:
+                try:
+                    output[name] = hex(int(value, 2)).upper()[2:]
+                    continue
+                except Exception as e:
+                    pass
+
             output[name] = value
     return output
 
@@ -102,9 +109,9 @@ class RecordBase:
         """Recursively pack TLB type to CellBuilder"""
         cb.store_slice_or_tlb(value)
 
-    def dump(self):
+    def dump(self, dump_bin_as_hex=True):
         """Recursively convert TLB to dict"""
-        return rec_dump(self)
+        return rec_dump(self, dump_bin_as_hex=dump_bin_as_hex)
 
     def to_dict(self, rec_unpack=False, convert_cells_to_bocs=False):
         answer = {}