+++ CHANGELOG.md
@@ -2,11 +2,16 @@
 
 - Smart-contract address class added
 - Mnemonic / (ED25519) PublicKey / (ED25519) PrivateKey classes added
-- Critical change: Run method in TVM will throw error on unsuccessful exit code
 - Merge testnet on commit:
 - Allow disassemble libs (root cell & PUSREF*)
 - Update C7 to "TVM Upgrade 2023.07", add BlockId
 - Add set_prev_blocks_info with BlockId's to emulator
+- Add shard_is_ancestor, shard_child to tonpy.utils.shard_account
+- Add BlockId / BlockIdExt python native support in C++, add wrappers
+- Allow VmDict initialization from PyDict
+- Add `combine_with` in VmDict to fast combine dictionaries
+- Critical change: TLB dump - forced `dump_bin_as_hex` to true. This means if bitsting is `x % 8 == 0` then it'll be dumped as hex (this is most of cases)
+- Critical change: Run method in TVM will throw error on unsuccessful exit code
 - Fix: emulator, account none several transaction emulation
 - Fix: C7 helper
 - Fix: Emulator set_libs